feat: refactor API dependencies and introduce composition for repository wiring
- Removed direct dependencies on AsyncSession and PostgreSQL repositories from API routers and dependencies.
- Introduced a new composition module to centralize repository wiring, enhancing modularity and adherence to hexagonal architecture.
- Updated API routers to utilize the new composition functions for obtaining repositories, improving code clarity and maintainability.
- Adjusted tests to reflect changes in dependency imports, ensuring proper functionality and coverage.

This refactor streamlines the dependency management across the application, promoting better separation of concerns and easier testing.
